fix tests and naming
