+++ haystack/preview/dataclasses/__init__.py
@@ -1,5 +1,5 @@
 from haystack.preview.dataclasses.document import Document
 from haystack.preview.dataclasses.answer import ExtractedAnswer, GeneratedAnswer, Answer
-from haystack.preview.dataclasses.blob import Blob
+from haystack.preview.dataclasses.byte_stream import ByteStream
 
-__all__ = ["Document", "ExtractedAnswer", "GeneratedAnswer", "Answer", "Blob"]
+__all__ = ["Document", "ExtractedAnswer", "GeneratedAnswer", "Answer", "ByteStream"]
